Modernize ansi-to-react and try out React.memo
Blocked by https://github.com/airbnb/enzyme/issues/1875 and https://github.com/DefinitelyTyped/DefinitelyTyped/pull/29990
