+++ applications/commuter/package.json
@@ -73,8 +73,8 @@
     "node-fetch": "^2.0.0",
     "nprogress": "^0.2.0",
     "prop-types": "^15.6.1",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-hot-loader": "^4.1.2",
     "s3-proxy": "^1.1.0",
     "util": "^0.11.0",
