+++ applications/jupyter-extension/nteract_on_jupyter/package.json
@@ -43,8 +43,8 @@
     "jquery": "^3.2.1",
     "moment": "^2.22.1",
     "monaco-editor-webpack-plugin": "^1.5.1",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-hot-loader": "^4.1.2",
     "react-notification-system": "^0.2.16",
     "react-redux": "^5.0.7",
