+++ applications/play/package.json
@@ -29,8 +29,8 @@
     "next": "7.0.2",
     "next-redux-wrapper": "^2.0.0",
     "object-path": "^0.11.4",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-redux": "^5.0.7",
     "redux": "^4.0.0",
     "redux-observable": "^1.0.0",
