+++ package.json
@@ -162,8 +162,8 @@
     "node-fetch": "2.1.2",
     "prettier": "^1.11.0",
     "raf": "^3.4.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-notification-system": "^0.2.16",
     "react-redux": "^5.0.7",
     "react-styleguidist": "^8.0.0",
