+++ packages/ansi-to-react/package.json
@@ -28,7 +28,7 @@
     "escape-carriage": "^1.3.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2"
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0"
   }
 }
