+++ packages/connected-components/package.json
@@ -38,7 +38,7 @@
   },
   "peerDependencies": {
     "immutable": "^4.0.0-rc.9",
-    "react": "^16.3.2",
+    "react": "^16.6.0",
     "styled-jsx": "^3.1.0"
   },
   "author": "Kyle Kelley <rgbkrk@gmail.com>",
