+++ packages/display-area/package.json
@@ -31,7 +31,7 @@
     "immutable": "^4.0.0-rc.9"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "author": "Kyle Kelley <rgbkrk@gmail.com>",
   "license": "BSD-3-Clause"
