+++ packages/dropdown-menu/package.json
@@ -31,8 +31,8 @@
     "styled-jsx": "^3.1.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2"
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0"
   },
   "author": "Kyle Kelley <rgbkrk@gmail.com>",
   "license": "BSD-3-Clause"
