+++ packages/jupyter-widgets/package.json
@@ -37,6 +37,6 @@
     "react-json-tree": "^0.11.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   }
 }
