+++ packages/notebook-render/package.json
@@ -36,7 +36,7 @@
   },
   "peerDependencies": {
     "immutable": "^4.0.0-rc.9",
-    "react": "^16.3.2",
+    "react": "^16.6.0",
     "styled-jsx": "^3.1.0"
   },
   "author": "Benjamin Abel <dev.abel@laposte.net>",
