+++ packages/octicons/package.json
@@ -19,7 +19,7 @@
     "babel-runtime": "^6.26.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "publishConfig": {
     "access": "public"
