+++ packages/outputs/package.json
@@ -19,7 +19,7 @@
     "ansi-to-react": "^3.3.5"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "publishConfig": {
     "access": "public"
