+++ packages/presentational-components/package.json
@@ -24,7 +24,7 @@
     "styled-jsx": "^3.1.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2",
+    "react": "^16.6.0",
     "styled-jsx": "^3.1.0"
   },
   "publishConfig": {
