+++ packages/timeago/package.json
@@ -36,7 +36,7 @@
   },
   "homepage": "https://github.com/nteract/nteract",
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "dependencies": {
     "@babel/runtime-corejs2": "^7.0.0",
