+++ packages/transform-geojson/package.json
@@ -21,7 +21,7 @@
     "leaflet": "^1.0.3"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "publishConfig": {
     "access": "public"
