+++ packages/transform-model-debug/package.json
@@ -24,7 +24,7 @@
     "babel-runtime": "^6.26.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "license": "BSD-3-Clause"
 }
