+++ packages/transform-plotly/package.json
@@ -26,7 +26,7 @@
     "lodash": "^4.17.4"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "license": "BSD-3-Clause"
 }
