+++ packages/transform-vdom/package.json
@@ -25,7 +25,7 @@
     "lodash": "^4.17.4"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "license": "BSD-3-Clause"
 }
