+++ packages/transforms-full/package.json
@@ -30,7 +30,7 @@
     "access": "public"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   },
   "dependencies": {
     "@babel/runtime-corejs2": "^7.0.0",
