+++ packages/transforms/package.json
@@ -38,6 +38,6 @@
     "react-json-tree": "^0.11.0"
   },
   "peerDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.6.0"
   }
 }
