+++ tsconfig.json
@@ -1,6 +1,8 @@
 {
   "compilerOptions": {
     "strict": true,
-    "declaration": true
+    "declaration": true,
+    "lib": ["esnext", "dom"],
+    "jsx": "react"
   }
 }
